requirements.txt
pyzotero>=1.5.0
sentence-transformers>=2.2.0
numpy>=1.24.0
scipy>=1.10.0
requests>=2.28.0
python-dotenv>=1.0.0
click>=8.1.0
tqdm>=4.65.0
